added footer for all components
